Overhaul hostapd management
Overhauls the management of hostapd in conjunction with wpa-supplicant. If we see wpa-supplicant changing the channel we now kill off hostapd, wait 10 seconds and then start up again with the latest observation of the wlan channel. While the hostapd script is somewhat more complex than what we had before, it has also eliminated complexity elsewhere.